requirements.txt
Adafruit-Blinka==6.15.0
adafruit-circuitpython-neopixel==6.2.2
adafruit-circuitpython-pixelbuf==1.1.0
Adafruit-PlatformDetect==3.17.1
Adafruit-PureIO==1.1.9
board==1.0
click==8.0.3
Flask==2.0.2
importlib-metadata==4.8.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pkg-resources==0.0.0
pyftdi==0.53.3
pyserial==3.5
python-chess==0.27.3
pyusb==1.2.1
rpi-ws281x==4.3.1
RPi.GPIO==0.7.0
sysv-ipc==1.1.0
typing-extensions==3.10.0.2
Werkzeug==2.0.2
zipp==3.6.0
